+++ CHANGELOG.md
@@ -8,9 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.27.1] - 2024-06-20
 
+### Added
+- Support for Claude 3.5 Sonnet in `AnthropicPromptDriver` and `AmazonBedrockPromptDriver`.
+
 ### Changed
 - Base Tool schema so that `input` is optional when no Tool Activity schema is set.
 - Tool Task system prompt for better results with lower-end models. 
+- Default Prompt Driver model to Claude 3.5 Sonnet in `AnthropicStructureConfig` and `AmazonBedrockStructureConfig.`
 
 ## [0.27.0] - 2024-06-19